requirements.txt
fastapi==0.105.0
fitz==0.0.1.dev2
gradio==4.11.0
img2table==1.2.6
ipython==8.12.3
jellyfish==1.0.3
matplotlib==3.7.4
numpy==1.24.4
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
paddlenlp==2.5.2
paddlepaddle==2.3.2
pandas==2.0.3
Pillow==10.1.0
Pillow==10.1.0
pytest==7.4.3
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
Shapely==2.0.2
Werkzeug==3.0.1
